utils/requirements.txt
requests
psycopg2-binary
autopep8
python-dotenv
sqlalchemy